feat(redux burger-constructor): выполнена типизация
