+++ src/components/burger-ingredients/ingredient-item/ingredient-item.tsx
@@ -6,7 +6,7 @@ import {
 import { useSelector } from 'react-redux';
 import { useDrag } from 'react-dnd';
 // @ts-expect-error: Could not find a declaration file for module '@services/burger-constructor/reducer'.
-import { getConstructorIngredients } from '@services/burger-constructor/reducer';
+import { getConstructorIngredients } from '@/services/burger-constructor/reducer';
 import { useLocation, Link } from 'react-router';
 import { TConstructorIngredients, TIngredient } from '@/utils/types';
